refactor (ChessGame) : 이름, 패키지 변경
