+++ src/main/java/chess/ChessApplication.java
@@ -1,6 +1,5 @@
 package chess;
 
-import chess.controller.ChessGameController;
 import chess.view.InputView;
 import chess.view.OutputView;
 
@@ -9,7 +8,7 @@ public class ChessApplication {
     public static void main(String[] args) {
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
-        ChessGameController chessGameController = new ChessGameController(inputView, outputView);
-        chessGameController.run();
+        ChessGame chessGame = new ChessGame(inputView, outputView);
+        chessGame.run();
     }
 }
